chore: add and apply integration suite FormatMixin
