+++ packages/field/test/field-integrations.test.js
@@ -1,6 +1,7 @@
 import { defineCE } from '@open-wc/testing';
 import { runInteractionStateMixinSuite } from '../test-suites/InteractionStateMixin.suite.js';
 import '../lion-field.js';
+import { runFormatMixinSuite } from '../test-suites/FormatMixin.suite.js';
 
 const fieldTagString = defineCE(
   class extends customElements.get('lion-field') {
@@ -18,5 +19,9 @@ describe('<lion-field> integrations', () => {
   runInteractionStateMixinSuite({
     tagString: fieldTagString,
     suffix: 'lion-field',
+  });
+
+  runFormatMixinSuite({
+    tagString: fieldTagString,
   });
 });
