+++ packages/input-amount/test/lion-input-amount-integrations.test.js
@@ -1,17 +1,18 @@
 import { runInteractionStateMixinSuite } from '@lion/field/test-suites/InteractionStateMixin.suite.js';
 import { runFormatMixinSuite } from '@lion/field/test-suites/FormatMixin.suite.js';
-import '../lion-input-datepicker.js';
+import '../lion-input-amount.js';
 
-const tagString = 'lion-input-datepicker';
-describe('<lion-input-datepicker> integrations', () => {
+const tagString = 'lion-input-amount';
+
+describe('<lion-input-amount> integrations', () => {
   runInteractionStateMixinSuite({
     tagString,
-    suffix: tagString,
-    allowedModelValueTypes: [Date],
+    suffix: 'lion-input-amount',
+    allowedModelValueTypes: [Number],
   });
 
   runFormatMixinSuite({
     tagString,
-    modelValueType: Date,
+    modelValueType: Number,
   });
 });
