+++ packages/input-date/test/lion-input-date-integrations.test.js
@@ -1,10 +1,18 @@
 import { runInteractionStateMixinSuite } from '@lion/field/test-suites/InteractionStateMixin.suite.js';
+import { runFormatMixinSuite } from '@lion/field/test-suites/FormatMixin.suite.js';
 import '../lion-input-date.js';
+
+const tagString = 'lion-input-date';
 
 describe('<lion-input-date> integrations', () => {
   runInteractionStateMixinSuite({
-    tagString: 'lion-input-date',
-    suffix: 'lion-input-date',
+    tagString,
+    suffix: tagString,
     allowedModelValueTypes: [Date],
+  });
+
+  runFormatMixinSuite({
+    tagString,
+    modelValueType: Date,
   });
 });
