+++ packages/input-email/test/lion-input-email-integrations.test.js
@@ -1,0 +1,11 @@
+import { runFormatMixinSuite } from '@lion/field/test-suites/FormatMixin.suite.js';
+import '../lion-input-email.js';
+
+const tagString = 'lion-input-email';
+
+describe('<lion-input-email> integrations', () => {
+  runFormatMixinSuite({
+    tagString,
+    modelValueType: 'email',
+  });
+});
